feat: Add SQL query to rank country origins of bands by number of fans
This commit adds a new SQL query in the file "2-fans.sql" that ranks the country origins of bands based on the number of fans. The query selects the "origin" column and calculates the total number of fans for each country using the "SUM" function. The results are then grouped by origin and ordered in descending order by the number of fans.