feat: Add MongoDB query to insert one object to 'school' collection in my_db
This commit adds a MongoDB query to insert one object to the 'school' collection in the 'my_db' database. The object being inserted has a 'name' field with the value "Holberton school". This enhancement improves the functionality of the code by allowing the insertion of a new document into the 'school' collection.