chore: Add function to retrieve all documents from MongoDB collection
This commit adds a new function `list_all` to retrieve all documents from a MongoDB collection. The function takes a `mongo_collection` parameter and returns a list of all documents in the collection. This enhancement improves the functionality of the code by providing a convenient way to retrieve all documents from a collection.