feat: Add insert_school function and update main script
This commit adds the `insert_school` function to the code, allowing the insertion of a new school document into the MongoDB collection. It also updates the main script to demonstrate the usage of the `insert_school` function and the `list_all` function to print the details of all schools in the collection. This enhancement improves the functionality of the code by providing the ability to insert and retrieve school documents from the database.